refactor: hook for handle follow status
